dropdown with traits shows correct data
