+++ src/components/FilterMenuNames.jsx
@@ -1,0 +1,34 @@
+import React, { useState, useEffect } from 'react';
+
+const FilterMenuNames = ({ content, changeOption }) => {
+    const [cardSelected, selectCard] = useState("");
+
+    const handleChange = (e) => {
+        const selectedValue = e.target.value;
+        selectCard(selectedValue);
+        changeOption(selectedValue);
+    };
+
+    return (
+
+        <>
+            <select id="name" value={cardSelected} onChange={handleChange}>
+                {
+                  content.map((item) => (
+                    <option key={item.id} value={item.name}
+                    >
+                      {
+                        item.traits.map(a => { 
+                          if (a['trait_type'] === "Legend")
+                            return a['value']
+                        }
+                    )}
+                    </option>
+                ))}
+            </select>
+            <p>{cardSelected}</p>
+        </>
+    )
+};
+
+export default FilterMenuNames;
